react component file restructure
