+++ client/debug.log
@@ -1,0 +1,4 @@
+[0826/205216.604:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0826/205216.605:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0826/205938.521:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0826/205938.522:ERROR:crash_report_database_win.cc(469)] failed to stat report
