+++ client/src/App.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import NavTabs from "./components/NavTabs";
-import Landing from "./components/pages/Landing";
-import Video from "./components/pages/Video";
-import Resources from "./components/pages/Resources";
-import Contact from "./components/pages/Contact";
-import Admin from "./components/pages/Admin";
-import Footer from "./components/Footer/Footer";
+import Landing from "./pages/jsx/Landing";
+import Video from "./pages/jsx/Video";
+import Resources from "./pages/jsx/Resources";
+import Contact from "./pages/jsx/Contact";
+import Admin from "./pages/jsx/Admin";
+import Footer from "./components/Footer";
 
 function App() {
   return (
