+++ client/src/components/Example/index.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import './style.css';
+
+
