+++ client/src/components/Footer/index.jsx
@@ -1,12 +1,13 @@
-//Dependency
-import React from "react";
+//DEPENDENCIES
+import React from 'react';
+import './style.css';
 
 //Create functional component
 function Footer() {
   return (
     <footer className="page-footer font-small blue">
       <div className="footer-copyright text-center py-3">
-        &copy;{new Date().getFullYear()} Copyright:Organize2.0
+        Organize 2.0 | &copy;{new Date().getFullYear()}
       </div>
     </footer>
   );
