--- client/src/pages/placeholder.txt
@@ -1,1 +1,0 @@
-placeholder
